Rename file, and add env var description.
