+++ torch_xla/csrc/aten_autograd_ops.cpp
@@ -5,7 +5,7 @@
 #include <ATen/native/CPUFallback.h>
 #include <c10/core/impl/PythonDispatcherTLS.h>
 
-#include "torch_xla/csrc/aten_cpu_fallback.h"
+#include "torch_xla/csrc/aten_fallback.h"
 #include "torch_xla/csrc/aten_xla_bridge.h"
 #include "torch_xla/csrc/helpers.h"
 #include "torch_xla/csrc/tensor_methods.h"
